converted sql queries to prisma

+++ src/store/actions/todo-actions/deleteTodoAction.js
@@ -2,9 +2,9 @@ import { createAsyncThunk } from '@reduxjs/toolkit'
 
 import API from '@/utils/axios'
 
-export const deleteTodo = createAsyncThunk('todos/deleteTodo', async todoId => {
+export const deleteTodo = createAsyncThunk('todos/deleteTodo', async ({ id: todoId, userId }) => {
   try {
-    await API.delete(`/api/todos/${todoId}`)
+    await API.delete(`/api/todos/${todoId}`, { data: { userId } })
     return todoId
   } catch (error) {
     throw new Error(error.message)
